completes v2.0.0

+++ controllers/movies.js
@@ -3,6 +3,14 @@ const jwt = require('jsonwebtoken');
 const Movie = require('../models/movie');
 const User = require('../models/user');
 
+const getTokenFrom = (req) => {
+  const authorization = req.get('authorization');
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7);
+  }
+  return null;
+};
+
 moviesRouter.get('/', async (_, res) => {
   const movies = await Movie.find({}).populate('user', {
     username: 1,
@@ -16,11 +24,9 @@ moviesRouter.post('/', async (req, res) => {
   const token = getTokenFrom(req);
   const decodedToken = jwt.verify(token, process.env.SECRET);
   if (!token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' });
+    return res.status(401).json({ error: 'token missing or invalid' });
   }
   const user = await User.findById(decodedToken.id);
-
-  user = await User.findById(body.userId);
 
   const movie = new Movie({
     title: body.title,
@@ -38,12 +44,27 @@ moviesRouter.post('/', async (req, res) => {
 });
 
 moviesRouter.delete('/:id', async (req, res) => {
+  const token = getTokenFrom(req);
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: 'token missing or invalid' });
+  }
+  let user = await User.findById(decodedToken.id);
+  user.movies = user.movies.filter(
+    (movie) => movie.toString() !== req.params.id.toString()
+  );
+  await user.save();
+
   await Movie.findByIdAndRemove(req.params.id);
   res.status(204).end();
 });
 
 moviesRouter.put('/:id', async (req, res) => {
   const body = req.body;
+  const token = getTokenFrom(req);
+  if (!token) {
+    return res.status(401).json({ error: 'token missing' });
+  }
 
   const movie = {
     likes: body.likes,
